+++ backend/server/restservice/machines_test.go
@@ -2827,7 +2827,7 @@ func TestDeleteDaemon(t *testing.T) {
 
 	_ = rapi.SessionManager.LoginHandler(ctx, &dbmodel.SystemUser{
 		ID: 42, Groups: []*dbmodel.SystemGroup{
-			{ID: int(authenticationcallouts.UserGroupIDSuperAdmin)},
+			{ID: int64(authenticationcallouts.UserGroupIDSuperAdmin)},
 		},
 	})
