+++ backend/server/restservice/sharednetworks_test.go
@@ -1042,7 +1042,7 @@ func TestCreateSharedNetwork4BeginSubmit(t *testing.T) {
 	}
 
 	// Make sure that the transaction is done.
-	cctx, _ := cm.RecoverContext(transactionID, int64(user.ID))
+	cctx, _ := cm.RecoverContext(transactionID, user.ID)
 	// Remove the context from the config manager before testing that
 	// the returned context is nil. If it happens to be non-nil the
 	// require.Nil() would otherwise spit out errors about the concurrent
@@ -1462,7 +1462,7 @@ func TestCreateSharedNetworkBeginCancel(t *testing.T) {
 	rsp2 := rapi.CreateSharedNetworkDelete(ctx, params2)
 	require.IsType(t, &dhcp.CreateSharedNetworkDeleteOK{}, rsp2)
 
-	cctx, _ := cm.RecoverContext(transactionID, int64(user.ID))
+	cctx, _ := cm.RecoverContext(transactionID, user.ID)
 	// Remove the context from the config manager before testing that
 	// the returned context is nil. If it happens to be non-nil the
 	// require.Nil() would otherwise spit out errors about the concurrent
@@ -1873,7 +1873,7 @@ func TestUpdateSharedNetwork4BeginSubmit(t *testing.T) {
 	}
 
 	// Make sure that the transaction is done.
-	cctx, _ := cm.RecoverContext(transactionID, int64(user.ID))
+	cctx, _ := cm.RecoverContext(transactionID, user.ID)
 	// Remove the context from the config manager before testing that
 	// the returned context is nil. If it happens to be non-nil the
 	// require.Nil() would otherwise spit out errors about the concurrent
@@ -2223,7 +2223,7 @@ func TestUpdateSharedNetwork6BeginSubmit(t *testing.T) {
 	}
 
 	// Make sure that the transaction is done.
-	cctx, _ := cm.RecoverContext(transactionID, int64(user.ID))
+	cctx, _ := cm.RecoverContext(transactionID, user.ID)
 	// Remove the context from the config manager before testing that
 	// the returned context is nil. If it happens to be non-nil the
 	// require.Nil() would otherwise spit out errors about the concurrent
@@ -2405,7 +2405,7 @@ func TestUpdateSharedNetworkBeginCancel(t *testing.T) {
 	rsp2 := rapi.UpdateSharedNetworkDelete(ctx, params2)
 	require.IsType(t, &dhcp.UpdateSharedNetworkDeleteOK{}, rsp2)
 
-	cctx, _ := cm.RecoverContext(transactionID, int64(user.ID))
+	cctx, _ := cm.RecoverContext(transactionID, user.ID)
 	// Remove the context from the config manager before testing that
 	// the returned context is nil. If it happens to be non-nil the
 	// require.Nil() would otherwise spit out errors about the concurrent
